+++ service/src/main/kotlin/io/provenance/explorer/grpc/extensions/Domain.kt
@@ -9,6 +9,7 @@ import cosmos.gov.v1beta1.Gov
 import cosmos.gov.v1beta1.Tx
 import cosmos.mint.v1beta1.Mint
 import cosmos.slashing.v1beta1.Slashing
+import cosmos.vesting.v1beta1.Vesting
 import io.grpc.Metadata
 import io.grpc.stub.AbstractStub
 import io.grpc.stub.MetadataUtils
@@ -16,14 +17,21 @@ import io.provenance.explorer.config.ExplorerProperties
 import io.provenance.explorer.config.ResourceNotFoundException
 import io.provenance.explorer.domain.core.logger
 import io.provenance.explorer.domain.core.toBech32Data
+import io.provenance.explorer.domain.extensions.diff
 import io.provenance.explorer.domain.extensions.edPubKeyToBech32
 import io.provenance.explorer.domain.extensions.secp256k1PubKeyToBech32
 import io.provenance.explorer.domain.extensions.secp256r1PubKeyToBech32
+import io.provenance.explorer.domain.extensions.toCoinStrList
+import io.provenance.explorer.domain.extensions.toDateTime
 import io.provenance.explorer.domain.extensions.toPercentage
 import io.provenance.explorer.domain.extensions.toSha256
+import io.provenance.explorer.domain.models.explorer.AccountVestingInfo
+import io.provenance.explorer.domain.models.explorer.CoinStr
 import io.provenance.explorer.domain.models.explorer.DistParams
 import io.provenance.explorer.domain.models.explorer.MintParams
 import io.provenance.explorer.domain.models.explorer.MsgBasedFee
+import io.provenance.explorer.domain.models.explorer.PeriodInSeconds
+import io.provenance.explorer.domain.models.explorer.PeriodicVestingInfo
 import io.provenance.explorer.domain.models.explorer.SlashingParams
 import io.provenance.explorer.domain.models.explorer.TallyingParams
 import io.provenance.explorer.domain.models.explorer.toData
@@ -32,6 +40,9 @@ import io.provenance.marker.v1.Access
 import io.provenance.marker.v1.MarkerAccount
 import io.provenance.marker.v1.MarkerStatus
 import io.provenance.msgfees.v1.MsgFee
+import org.joda.time.DateTime
+
+//region GRPC query
 
 const val BLOCK_HEIGHT = "x-cosmos-block-height"
 
@@ -40,16 +51,35 @@ fun <S : AbstractStub<S>> S.addBlockHeightToQuery(blockHeight: String): S =
         .also { it.put(Metadata.Key.of(BLOCK_HEIGHT, Metadata.ASCII_STRING_MARSHALLER), blockHeight) }
         .let { MetadataUtils.attachHeaders(this, it) }
 
-// Marker Extensions
+fun getPaginationBuilder(offset: Int, limit: Int) =
+    Pagination.PageRequest.newBuilder().setOffset(offset.toLong()).setLimit(limit.toLong()).setCountTotal(true)
+
+fun getPaginationBuilderNoCount(offset: Int, limit: Int) =
+    Pagination.PageRequest.newBuilder().setOffset(offset.toLong()).setLimit(limit.toLong())
+
+fun getPagination(offset: Int, limit: Int) =
+    pageRequest {
+        this.offset = offset.toLong()
+        this.limit = limit.toLong()
+        this.countTotal = true
+    }
+
+fun getPaginationNoCount(offset: Int, limit: Int) =
+    pageRequest {
+        this.offset = offset.toLong()
+        this.limit = limit.toLong()
+    }
+
+//endregion
+
+//region Marker Extensions
 fun String.getTypeShortName() = this.split(".").last()
 
 fun Any.toMarker(): MarkerAccount =
-    this.typeUrl.getTypeShortName().let {
-        when (it) {
-            MarkerAccount::class.java.simpleName -> this.unpack(MarkerAccount::class.java)
-            else -> {
-                throw ResourceNotFoundException("This marker type has not been mapped yet")
-            }
+    when (this.typeUrl.getTypeShortName()) {
+        MarkerAccount::class.java.simpleName -> this.unpack(MarkerAccount::class.java)
+        else -> {
+            throw ResourceNotFoundException("This marker type has not been mapped yet")
         }
     }
 
@@ -74,7 +104,9 @@ fun MarkerAccount.getManagingAccounts(): MutableMap<String, List<String>> {
     }
 }
 
-// Account Extensions
+//endregion
+
+//region Account Extensions
 fun Any.getModuleAccName() =
     if (this.typeUrl.getTypeShortName() == Auth.ModuleAccount::class.java.simpleName)
         this.unpack(Auth.ModuleAccount::class.java).name
@@ -98,6 +130,105 @@ fun String.isStandardAddress(props: ExplorerProperties) =
     this.startsWith(props.provAccPrefix()) && !this.startsWith(props.provValOperPrefix())
 fun String.isValidatorAddress(props: ExplorerProperties) = this.startsWith(props.provValOperPrefix())
 
+// Ref https://docs.cosmos.network/master/modules/auth/05_vesting.html#continuousvestingaccount
+// for info on how the periods are calced
+fun Any.toVestingData(initialDate: DateTime?, continuousPeriod: PeriodInSeconds): AccountVestingInfo {
+    val now = DateTime.now().millis / 1000
+    when (this.typeUrl.getTypeShortName()) {
+        Vesting.ContinuousVestingAccount::class.java.simpleName ->
+            // Given the PeriodInSeconds, chunk the totalTime and calculate how much coin will be vested at the end
+            // of the given period. Continuous technically vests every second, but that can be cumbersome to fetch.
+            this.unpack(Vesting.ContinuousVestingAccount::class.java).let { acc ->
+                val totalTime = acc.baseVestingAccount.endTime - acc.startTime
+                var runningTime = acc.startTime // returns the actual vestingTime
+                var prevCoins = emptyList<CoinStr>() // reset for every period with the latest percentage amounts
+                val periods = (acc.startTime until acc.baseVestingAccount.endTime)
+                    .chunked(continuousPeriod.seconds)
+                    .mapNotNull { list ->
+                        if (list.last() == acc.startTime) return@mapNotNull null
+                        runningTime += list.size // Updated to show the actual vestingTime for the period
+                        val elapsedTime = runningTime - acc.startTime // elapsed time up to the vestingTime
+                        // Calcs the total percentage up to the current vestingTime of vested coins
+                        val newCoin = acc.baseVestingAccount.originalVestingList.map { it.toPercentage(elapsedTime, totalTime) }
+
+                        PeriodicVestingInfo(
+                            list.size.toLong(), // How long the period is in seconds
+                            prevCoins.diff(newCoin), // diffs the old percentages with the new percentages to get the period's values
+                            runningTime.toDateTime(), // vestingTime for the period
+                            runningTime <= now // compared to NOW, is it vested
+                        ).also { prevCoins = newCoin }
+                    }
+                return AccountVestingInfo(
+                    now.toDateTime(),
+                    acc.baseVestingAccount.endTime.toDateTime(),
+                    acc.baseVestingAccount.originalVestingList.toCoinStrList(),
+                    acc.startTime.toDateTime(),
+                    periods
+                )
+            }
+        Vesting.DelayedVestingAccount::class.java.simpleName ->
+            // Delayed is a one and done period, timed to the endTime
+            this.unpack(Vesting.DelayedVestingAccount::class.java).let {
+                val period = PeriodicVestingInfo(
+                    0L,
+                    it.baseVestingAccount.originalVestingList.toCoinStrList(),
+                    it.baseVestingAccount.endTime.toDateTime(),
+                    it.baseVestingAccount.endTime < now
+                )
+                return AccountVestingInfo(
+                    now.toDateTime(),
+                    it.baseVestingAccount.endTime.toDateTime(),
+                    it.baseVestingAccount.originalVestingList.toCoinStrList(),
+                    initialDate ?: now.toDateTime(),
+                    listOf(period)
+                )
+            }
+        Vesting.PeriodicVestingAccount::class.java.simpleName ->
+            // Periodic repackages the given Period objects
+            this.unpack(Vesting.PeriodicVestingAccount::class.java).let { acc ->
+                var runningTime = acc.startTime
+                val periods = acc.vestingPeriodsList.map {
+                    runningTime += it.length
+                    PeriodicVestingInfo(
+                        it.length,
+                        it.amountList.toCoinStrList(),
+                        runningTime.toDateTime(),
+                        runningTime < now
+                    )
+                }
+
+                return AccountVestingInfo(
+                    now.toDateTime(),
+                    acc.baseVestingAccount.endTime.toDateTime(),
+                    acc.baseVestingAccount.originalVestingList.toCoinStrList(),
+                    acc.startTime.toDateTime(),
+                    periods
+                )
+            }
+        Vesting.PermanentLockedAccount::class.java.simpleName ->
+            this.unpack(Vesting.PermanentLockedAccount::class.java).let {
+                return AccountVestingInfo(
+                    now.toDateTime(),
+                    it.baseVestingAccount.endTime.toDateTime(),
+                    it.baseVestingAccount.originalVestingList.toCoinStrList(),
+                    initialDate ?: now.toDateTime()
+                )
+            }
+
+        else -> throw ResourceNotFoundException("This Vesting Account type has not been mapped: ${this.typeUrl.getTypeShortName()}")
+    }
+}
+
+fun Any.isVesting() =
+    this.typeUrl.getTypeShortName() == Vesting.ContinuousVestingAccount::class.java.simpleName ||
+        this.typeUrl.getTypeShortName() == Vesting.DelayedVestingAccount::class.java.simpleName ||
+        this.typeUrl.getTypeShortName() == Vesting.PeriodicVestingAccount::class.java.simpleName ||
+        this.typeUrl.getTypeShortName() == Vesting.PermanentLockedAccount::class.java.simpleName
+
+//endregion
+
+//region IBC
+
 // TODO: Once cosmos-sdk implements a grpc endpoint for this we can replace this with grpc Issue: https://github.com/cosmos/cosmos-sdk/issues/9437
 fun getEscrowAccountAddress(portId: String, channelId: String, hrpPrefix: String): String {
     val contents = "$portId/$channelId".toByteArray()
@@ -106,24 +237,9 @@ fun getEscrowAccountAddress(portId: String, channelId: String, hrpPrefix: String
     return hash.toBech32Data(hrpPrefix).address
 }
 
-fun getPaginationBuilder(offset: Int, limit: Int) =
-    Pagination.PageRequest.newBuilder().setOffset(offset.toLong()).setLimit(limit.toLong()).setCountTotal(true)
+//endregion
 
-fun getPaginationBuilderNoCount(offset: Int, limit: Int) =
-    Pagination.PageRequest.newBuilder().setOffset(offset.toLong()).setLimit(limit.toLong())
-
-fun getPagination(offset: Int, limit: Int) =
-    pageRequest {
-        this.offset = offset.toLong()
-        this.limit = limit.toLong()
-        this.countTotal = true
-    }
-
-fun getPaginationNoCount(offset: Int, limit: Int) =
-    pageRequest {
-        this.offset = offset.toLong()
-        this.limit = limit.toLong()
-    }
+//region To DTOs
 
 fun Distribution.Params.toDto() = DistParams(
     this.communityTax.toPercentage(),
@@ -159,3 +275,5 @@ fun MsgFee.toDto() = MsgBasedFee(this.msgTypeUrl, this.additionalFee.toData())
 
 fun Tx.MsgVote.toWeightedVote() =
     Gov.WeightedVoteOption.newBuilder().setOption(this.option).setWeight("1000000000000000000").build()
+
+//endregion